Keep `CVersion` in the `vminitd` package
